+++ VacanzeApi/Services/Controllers/Grupo5/AutoController.cs
@@ -14,7 +14,7 @@ namespace vacanze_back.VacanzeApi.Services.Controllers.Grupo5
     {
         /*https://localhost:5001/api/Auto/agregar/toyotas/corola/4/true/ac366df/23000/<aqui_va_la_foto>/1*/
         [HttpGet("agregar/{make}/{model}/{capacity}/{status}/{licence}/{price}/{picture}/{place}")]
-        public ActionResult<IEnumerable<String>> agregar(string make,string model,int capacity,bool status, string licence, int price, string picture , int place)
+        public ActionResult<IEnumerable<String>> agregar(string make,string model,int capacity,bool status, string licence, float price, string picture , int place)
         {
             ConnectAuto conec= new ConnectAuto();
             Auto auto = new Auto(make,model,capacity,status,licence,price,picture,place);
@@ -51,7 +51,7 @@ namespace vacanze_back.VacanzeApi.Services.Controllers.Grupo5
         }
 		/*https://localhost:5001/api/Auto/modificar/2/cambio/cambio/532/false/cambio/12345/cambio/2 */
         [HttpGet("modificar/{id}/{make}/{model}/{capacity}/{status}/{licence}/{price}/{picture}/{place}")]
-        public  IActionResult  modify(int id, string make,string model,int capacity,bool status, string licence, int price, string picture , int place)
+        public  IActionResult  modify(int id, string make,string model,int capacity,bool status, string licence, float price, string picture , int place)
         {
             try{
                 ConnectAuto conec= new ConnectAuto();
